+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Issue with changes to ingress and reports
 ### Changed
 - Better help text on cli
+- Adding variants to csv output
 - Keep only mapped reads
 ### Fixed
 - Downgraded pomoxis to prevent issues with empty bams